% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_secret.R
\name{add_secret}
\alias{add_secret}
\title{add_secret}
\usage{
add_secret(secret, force_update = FALSE)
}
\arguments{
\item{secret}{A character argument containing an secret generated from the XNAT website}

\item{force_update}{This boolean value provides an option to skip the utils::menu selection, Default: FALSE}
}
\value{
This function does not return any values.
}
\description{
This is a convenience function to help during the set up of the
kuadrc.xnat package. A user can use this function to help add or update their
secret as a variable in their computer's R environment.
}
\details{
This is a convenience function to help during the set up of the
kuadrc.xnat package. A user can use this function to help add or update their
secret as a variable in their computer's R environment.
}
\seealso{
\code{\link[utils]{menu}}
}
