% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_server.R
\name{add_server}
\alias{add_server}
\title{add_server}
\usage{
add_server(server, force_update = FALSE)
}
\arguments{
\item{server}{A character argument containing a URL starting with "https://"
followed by the university's XNAT server (e.g., "https://xnat.university.edu")}

\item{force_update}{This boolean value provides an option to skip the utils::menu selection, Default: FALSE}
}
\value{
This function does not return any values.
}
\description{
This is a convenience function to help during the set up of the
kuadrc.xnat package. A user can use this function to help add their university
XNAT server as a variable in their computer's R environment.
}
\details{
This is a convenience function to help during the set up of the
kuadrc.xnat package. A user can use this function to help add their university
XNAT server as a variable in their computer's R environment.
}
\seealso{
\code{\link[utils]{menu}}
}
