% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_url.R
\name{construct_url}
\alias{construct_url}
\title{construct_url}
\usage{
construct_url(server, projects = NULL, subjects = NULL, experiments = NULL)
}
\arguments{
\item{server}{A character argument containing a URL starting with "https://"
followed by the university's XNAT server (e.g., "https://xnat.university.edu")}

\item{projects}{A project number in which a user has access, Default: NULL}

\item{subjects}{A subject identifier, Default: NULL}

\item{experiments}{A numeric value corresponding to the subject's experiment or session, Default: NULL}
}
\value{
A character value containing a formatted URL to the XNAT API endpoint
}
\description{
Helps construct the URL to retrieve projects, subjects, experiments,
and scans from the XNAT API endpoints
}
\details{
Helps construct the URL to retrieve projects, subjects, experiments,
and scans from the XNAT API endpoints
}
