% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_nifti.R
\name{convert_to_nifti}
\alias{convert_to_nifti}
\title{convert_to_nifti}
\usage{
convert_to_nifti(directory)
}
\arguments{
\item{directory}{The location of the scans folder as downloaded with the `xnat_download` function.}
}
\value{
A character value containing the location of the nifti files
}
\description{
Reads in .dcm files and converts them to .nii
}
\details{
Reads in .dcm files and converts them to .nii
}
\seealso{
\code{\link[oro.dicom]{readDICOM}}, \code{\link[oro.dicom]{dicom2nifti}}
 \code{\link[oro.nifti]{writeNIfTI-methods}}
}
