% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{get_projects}
\usage{
get_projects(name = NULL, ...)
}
\arguments{
\item{name}{An optional name can be provided to only return the project number, Default: NULL}

\item{...}{Optional parameters that can include updated server, alias, and secret values that
will be added to a user's R environment. A user can also provide their username and password
as arguments to interact with the API without using an alias and secret.}
}
\value{
A data set of the user's approved projects or a character value if name
uniquely matches a study label in the data set. If name matches more than one
value in the data set, this function will return all values.
}
\description{
Retrieves a user's projects from the API endpoint
}
\details{
Retrieves a user's projects from the API endpoint
}
