% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scans.R
\name{get_scans}
\alias{get_scans}
\title{get_scans}
\usage{
get_scans(project = NULL, subject = NULL, experiment = NULL, scan = NULL, ...)
}
\arguments{
\item{project}{An optional project number can be provided to only return the experiments within that specific project, Default: NULL}

\item{subject}{An optional subject number can be provided to only return the experiments for a certain subject, Default: NULL}

\item{experiment}{An optional experiment number can be provided to only return those experiments, Default: NULL}

\item{scan}{An optional scan can be provided to only return those scans, Default: NULL}

\item{...}{Optional parameters that can include updated server, alias, and secret values that
will be added to a user's R environment. A user can also provide their username and password
as arguments to interact with the API without using an alias and secret.}
}
\value{
A data set of the user's approved scans or a character value if scan
uniquely matches an scan label in the data set. If name matches more than one
value in the data set, this function will return all values.
}
\description{
Retrieves all scans or scans from specific projects, subjects, and experiments
}
\details{
Retrieves all scans or scans from specific projects, subjects, and experiments
}
