% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subjects.R
\name{get_subjects}
\alias{get_subjects}
\title{get_subjects}
\usage{
get_subjects(project = NULL, subject = NULL, ...)
}
\arguments{
\item{project}{An optional project number can be provided to only return the experiments within that specific project, Default: NULL}

\item{subject}{An optional subject number can be provided to only return information from a certain subject, Default: NULL}

\item{...}{Optional parameters that can include updated server, alias, and secret values that
will be added to a user's R environment. A user can also provide their username and password
as arguments to interact with the API without using an alias and secret.}
}
\value{
A data set of the user's approved experiments or a character value if experiment
uniquely matches an experiment label in the data set. If name matches more than one
value in the data set, this function will return all values.
}
\description{
Retrieves all subjects or subjects from specific projects
}
\details{
Retrieves all subjects or subjects from specific projects
}
