% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuadrc.xnat-package.R
\docType{package}
\name{kuadrc.xnat-package}
\alias{kuadrc.xnat-package}
\title{kuadrc.xnat: Makes it easier to work with the XNAT API}
\description{
Eases the implementation of working with the XNAT API in R and offers ways to read in dicm file formats convert them to images that can be used on reports.
}
\author{
\strong{Maintainer}: Brian C. Helsel \email{bhelsel@kumc.edu}

}
\keyword{internal}
