% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nifti.R
\name{plot_nifti}
\alias{plot_nifti}
\title{plot_nifti}
\usage{
plot_nifti(
  nifti_data,
  plane = c("axial", "sagittal", "coronal"),
  index = NULL,
  n_gray = 64,
  adjust_brightness = 1
)
}
\arguments{
\item{nifti_data}{Data set from a .nii file read in by the `read_nifti` function}

\item{plane}{The plane to be displayed, Default: c("axial", "sagittal", "coronal")}

\item{index}{The slice to be displayed, Default: NULL}

\item{n_gray}{Number of grayscale colors to display, Default: 64}

\item{adjust_brightness}{Manually adjust brightness by multiplying by this factor, Default: 1}
}
\value{
A ggplot object of the brain image and its center
}
\description{
Uses ggplot2 to produce a plot containing the .nii image and center of mass
}
\details{
Uses ggplot2 to produce a plot containing the .nii image and center of mass
}
\seealso{
\code{\link[oro.nifti]{cal_min-methods}}, \code{\link[oro.nifti]{cal_max-methods}}
 \code{\link[ggplot2]{coord_fixed}}, \code{\link[ggplot2]{element}}
}
