% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_nifti_image.R
\name{save_nifti_image}
\alias{save_nifti_image}
\title{save_nifti_image}
\usage{
save_nifti_image(
  path = NULL,
  image_type = "png",
  width = 8.5,
  height = 11,
  dpi = 300,
  plane = c("", "axial", "sagittal", "coronal"),
  save_file_as = NULL,
  ...
)
}
\arguments{
\item{path}{The location of a nifti file or folder as created with the `convert_to_nifti` function.}

\item{image_type}{The type of image to include as the extension, Default: 'png'}

\item{width}{Plot width in inches or other units passed to `...`, Default: 8.5}

\item{height}{Plot height in inches or other units passed to `...`, Default: 11}

\item{dpi}{Plot resolution, Default: 300}

\item{plane}{The plane to be displayed, Default: c("axial", "sagittal", "coronal")}

\item{save_file_as}{An option to change the file to a name other than what was downloaded from xnat}

\item{...}{Other arguments passed on to the `plot_nifti` function}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Saves images after plotting with the `plot_nifti` function. This is
    convenience function to loop through all the .nii images in folder to plot and save
    the images as `image_type`.
}
\details{
Saves images after plotting with the `plot_nifti` function. This is
    convenience function to loop through all the .nii images in folder to plot and save
    the images as `image_type`. It uses the same folder structure as retrieved from
    the `xnat_download` function.
}
\seealso{
\code{\link[ggplot2]{ggsave}}
}
