% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_credentials.R
\name{validate_credentials}
\alias{validate_credentials}
\title{validate_credentials}
\usage{
validate_credentials(...)
}
\arguments{
\item{...}{Optional arguments of server, alias, and secret can be added. This
function will validate credentials based on the server, alias, and secret
stored in the R environment file if these parameters are not added.}
}
\value{
A list containing the server and user's validated access credentials.
}
\description{
Search for and validate the user's XNAT access token credentials.
}
\details{
Search for and validate the user's XNAT access token credentials.
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{authenticate}}, \code{\link[httr]{BROWSE}}, \code{\link[httr]{content}}, \code{\link[httr]{config}}
 \code{\link[utils]{menu}}
}
