% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_server.R
\name{validate_server}
\alias{validate_server}
\title{validate_credentials}
\usage{
validate_server(...)
}
\arguments{
\item{...}{Optional arguments of server can be added. This
function will validate credentials based on the server
stored in the R environment file if these parameters are not added.}
}
\value{
The server entered by the user or stored in the user's R environment.
}
\description{
Search for and validate the user's XNAT server.
}
\details{
Search for and validate the user's XNAT server.
}
