% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnat_download.R
\name{xnat_download}
\alias{xnat_download}
\title{xnat_download}
\usage{
xnat_download(outdir, ...)
}
\arguments{
\item{outdir}{A location to download the scan to in the user's directory.}

\item{...}{The user can optionally pass their username and password directly to the `xnat_download`
function instead of setting up an alias and secret. If `username` and `password` are not arguments
in the `xnat_download` function then the user will be prompted to create credentials.}
}
\value{
A character value containing the location of the downloads
}
\description{
Downloads scans from the XNAT image scan's API endpoint. A user can
interact with this function directly by adding their username and password or
alias and secret.
}
\details{
Downloads scans from the XNAT image scan's API endpoint. A user can
interact with this function directly by adding their username and password or
alias and secret.
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{authenticate}}, \code{\link[httr]{http_error}}, \code{\link[httr]{status_code}}, \code{\link[httr]{content}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
}
